+++ src/tcode_api/api/compat.py
@@ -326,6 +326,10 @@ tcode_api_compat_context = CompatContext(
                 "WellPlateDescription": 4,
                 "WellPlateDescriptor": 4,
             },
+            "v1.43.0": {
+                "CANMERA_CONNECT_WIFI": 1,
+                "CANMERA_NET_STATUS": 1,
+            },
         },
         migrations={
             "v1.37.0": {