+++ src/tcode_api/schemas/commands/__init__.py
@@ -13,6 +13,8 @@ from .calibrate_labware_height.latest import CALIBRATE_LABWARE_HEIGHT
 from .calibrate_labware_holder.latest import CALIBRATE_LABWARE_HOLDER
 from .calibrate_labware_well_depth.latest import CALIBRATE_LABWARE_WELL_DEPTH
 from .calibrate_tool.latest import CALIBRATE_TOOL
+from .canmera_connect_wifi.latest import CANMERA_CONNECT_WIFI
+from .canmera_net_status.latest import CANMERA_NET_STATUS
 from .comment.latest import COMMENT
 from .create_labware.latest import CREATE_LABWARE
 from .delete_labware.latest import DELETE_LABWARE
@@ -47,6 +49,8 @@ __all__ = [
     "CALIBRATE_LABWARE_HOLDER",
     "CALIBRATE_LABWARE_WELL_DEPTH",
     "CALIBRATE_TOOL",
+    "CANMERA_CONNECT_WIFI",
+    "CANMERA_NET_STATUS",
     "COMMENT",
     "CREATE_LABWARE",
     "DELETE_LABWARE",