+++ src/tcode_api/schemas/commands/union.py
@@ -19,6 +19,12 @@ from .calibrate_labware_well_depth.latest import (
 from .calibrate_tool.latest import (
     CALIBRATE_TOOL,
 )
+from .canmera_connect_wifi.latest import (
+    CANMERA_CONNECT_WIFI,
+)
+from .canmera_net_status.latest import (
+    CANMERA_NET_STATUS,
+)
 from .comment.latest import COMMENT
 from .create_labware.latest import CREATE_LABWARE
 from .delete_labware.latest import DELETE_LABWARE
@@ -58,6 +64,8 @@ TCode = Annotated[
     | CALIBRATE_LABWARE_HEIGHT
     | CALIBRATE_LABWARE_HOLDER
     | CALIBRATE_TOOL
+    | CANMERA_CONNECT_WIFI
+    | CANMERA_NET_STATUS
     | COMMENT
     | CREATE_LABWARE
     | DELETE_LABWARE